Remove factory function for creating subagents
