+++ src/agent/harness/subagents/geo.py
@@ -1,3 +1,9 @@
+import json
+
+from langchain.tools import ToolRuntime, tool
+from langchain_core.messages import ToolMessage
+from langgraph.types import Command
+
 from src.agent.harness.state import AoiRef
 
 _FIXTURES: dict[str, list[AoiRef]] = {
@@ -39,6 +45,8 @@ _FIXTURES: dict[str, list[AoiRef]] = {
 class GeoAgent:
     """Stub geo subagent. Resolves a place query to AoiRefs.
 
+    Will be replaced by a DSPy agent.
+
     Returns hard-coded fixtures for known names; otherwise a single
     deterministic ref so downstream tools have something to work with.
     """
@@ -57,3 +65,20 @@ class GeoAgent:
                 "subtype": None,
             }
         ]
+
+
+@tool
+async def geo_subagent(query: str, runtime: ToolRuntime) -> Command:
+    """Resolve a place-name query (e.g. "Para, Brazil", "neighbours of
+    Odisha", "1km buffer around Yellowstone") to one or more AOI refs.
+    Returns [{name, source, src_id, subtype}]. Updates state.aoi_refs."""
+    refs = await GeoAgent().resolve(query)
+    ref_dicts = [dict(r) for r in refs]
+    runtime.stream_writer({"type": "aoi_resolved", "aoi_refs": ref_dicts})
+    return Command(update={
+        "aoi_refs": refs,
+        "messages": [ToolMessage(
+            content=json.dumps(ref_dicts),
+            tool_call_id=runtime.tool_call_id,
+        )],
+    })
